Refactor highlight text function into utils module
